refs and config files created
