+++ src/javascript/refs.js
@@ -1,0 +1,5 @@
+export const refs = {
+  searchFormElement: document.querySelector('#search-form'),
+  searchInputElement: document.querySelector('#search-input'),
+  main: document.querySelector('#main'),
+};
